task2_NER_imageClassification/requirements.txt
torch==2.6.0
transformers
datasets
pandas
matplotlib
seaborn
Pillow
torchvision
scipy
numpy
argparse
jupyter